R200 second pass: "there is nothing clean left to cut" was a feeling, so it was measured
Every section of js/app-body.js (its own banner comments are the boundaries) went through
the surface tool and was ranked by (in + 2·out)/lines. The top four are CLEANER cuts than
anything taken in the first pass — so the feeling was simply wrong:

  js/i18n-late.js           107 lines  in 2  out 0   the late i18n keys + the ticker's settings
  js/wheel-zoom.js           66 lines  in 2  out 0   the wheel, and how fast the map answers it
  js/keyboard-shortcuts.js   76 lines  in 6  out 0   the keyboard, and the card that lists it
  js/label-occlusion.js      99 lines  in 5  out 1   names on top + the far-side markers

js/app-body.js 4,685 → 4,360 (5,149 at the start of the round, 5,375 before #R198).
341 of the 348 moved lines are byte-identical and the kept side is 4,337/4,337 identical
apart from the handle line and the two additions the cut required: the updateOcclusion
hoisting shim (same reason as the first pass — IM_HOST publishes it and the map wiring calls
it on every move) and a `get markersArray()` accessor, which must be live because
clearMarkers() REPLACES that array. Round total: 852 lines moved, 831 byte-identical, ten
modules.

One of the four writes host state — the `t` shortcut cycles light → dark → auto — through
IM_HOST's accessor pair, and tests/r165-checks.test.mjs names it as an owner.

Two checks had to be taught where things live now, both #R199's shape:
 · ⚠ scripts/static-checks.mjs said js/tile-warm.js "defines a factory nothing ever calls".
   The mount had not gone anywhere — it travelled with the block that always held it. The
   check's "page" (index.html + src/main.js + app-body + geo-engine) now also counts the js/
   files app-body statically imports, because that is what they are.
 · ⚠ r163 #2 failed on the new IM_HOST accessor until every line of its comment began with
   `*` — a bare continuation line reads as a member. Same trap #R198 hit.

Seven assertions re-pointed rather than relaxed. Full local suite green again: 591 node
checks, the browser plan's main pool, and the solo pool 31/31.
