+++ DEV-NOTES.md
@@ -30,8 +30,8 @@
 
 | | 前 | 後 |
 |---|---|---|
-| `js/app-body.js` | 5,149 | **4,685**（−464行・#R198 比 −690行） |
-| 出した本物の ES モジュール | 7本（#R199） | **+6本＝13本** |
+| `js/app-body.js` | 5,149 | **4,360**（−789行・#R198 比 −1,015行） |
+| 出した本物の ES モジュール | 7本（#R199） | **+10本＝17本** |
 | 夜側の最深部の暗さ（z1.2・対日点・海上） | 0.57（**計算値**：`1−(1−0.156)^5`） | **0.87（実測**：輝度 191.6 → 24.7） |
 | 時間旅行で太陽が動くか | **動かない**（2つの独立した理由で） | 動く（地平線 `#0a1526` ⇄ `#c9dcf0`） |
 
@@ -77,7 +77,34 @@
 生きたアクセサ経由（`currentMode`→`HOST.mode`・`currentMapType`→`HOST.mapType`・`terrain3D`・
 `currentLang`→`HOST.lang`・`countryDataLoaded`）。**残った側も 4,645/4,645行が同一**——
 ただし②で足した**巻き上げ shim の17行だけは純粋な追加**で、これを除いての一致（機械で取った）。
-6本のうち**HOST に書き戻すものは1つも無い**——そう選んだ（`tests/r200-checks ③` が検査する）。
+最初の6本のうち**HOST に書き戻すものは1つも無い**——そう選んだ（`tests/r200-checks ③` が検査する）。
+
+#### 1b. そして**全セクションを掃いた**——「もう切れる塊は無い」は測ってから言う
+
+CI が動かない待ち時間に、`js/app-body.js` の**全セクション**（既存のバナーコメントを境界にした）を
+1つずつ面測定にかけ、`(in + 2·out) / 行数` で並べた。⚠ **上位4つは、1回目に採った6本より綺麗だった**
+——つまり「良い切り口はもう無い」という感触のほうが間違っていた。#R169 が index.html で書いた
+「切れる塊が無い＝切り口を変える合図」の、測定版である。
+
+| 追加で採った候補 | 行 | in | out | 出したファイル |
+|---|---|---|---|---|
+| 遅れて足された i18n キー（5言語）＋ティッカー設定 | 107 | 2 | **0** | `js/i18n-late.js` |
+| ホイールズームと操作感度 | 66 | 2 | **0** | `js/wheel-zoom.js` |
+| キーボードショートカットと `?` の一覧 | 76 | 6 | **0** | `js/keyboard-shortcuts.js` |
+| 地名/国境を最前面に＋球の裏側のマーカーを隠す | 99 | 5 | 1 | `js/label-occlusion.js` |
+
+2回目の移送も機械で証明した：**348行のうち341行がバイト一致**、残った側は **4,337/4,337行が同一**
+（例外は `_IM_LABELOCCLUSION` を足したハンドル宣言の1行と、②と同じ理由の `updateOcclusion` の
+巻き上げ shim・`get markersArray()` アクセサ）。**ラウンド合計で852行・うち831行がバイト一致**。
+
+⚠ **`markersArray` のアクセサを足すとき `r163 #2` が落ちた**——IM_HOST 内のコメントは**全行を `*` で
+始める**必要がある（#R198 が同じ所で踏んだ罠。継続行が生テキストだと、メンバー列挙のパーサが
+それをメンバーだと思う）。⚠ **`scripts/static-checks.mjs` にも #R199 と同じ穴があった**：
+「`js/tile-warm.js` はファクトリを定義しているが誰も呼んでいない」と赤くなった。呼び出しは消えて
+いない——それが乗っていた**ブロックごと `js/label-occlusion.js` に移った**だけで、この検査の
+「ページ＝index.html＋src/main.js＋app-body＋geo-engine」という定義が古かった。**app-body が静的に
+import している兄弟ファイルもページの一部**として数えるようにした（#R199 が到達可能性の側に
+教えたのと同じ訂正）。
 
 ### ② ⚠ **自分で回帰を作り、ブラウザだけがそれを見つけた**——const は「同じ名前」ではない
 
@@ -172,11 +199,14 @@ _IM_LFAVS = makeLayerFavs(IM_HOST, { … });                                    
 
 - **新規**: `tests/r200-checks.test.mjs`（10件・両ファイルから導出）＋ `tests/r200.spec.js`（3件・
   レンダラでしか答えられないもの）。`npm run test:checks` に登録済み。
-- **移動で落ちた8件は、緩めずに指し先を直した**（#R199 §4 と同じ）：`r170`（既定タブ）・
+- **移動で落ちた8＋7件は、緩めずに指し先を直した**（#R199 §4 と同じ）：`r170`（既定タブ）・
   `r186`/`r188`/`r189`×2/`r190`（既定レイヤーと `defv` 世代印）・`r195`×2（セッションキーと
   サイドバー状態）を **`js/session-tabs.js` に直接**訊く。連結（`appShell()`）を検索するより厳しい
-  ——そのファイルに無ければ落ちる。⚠ 6本は `appShell()` に**足していない**（足すと「中心部が
-  小さくなった」という測定そのものが無意味になる）。
+  ——そのファイルに無ければ落ちる。⚠ 10本とも `appShell()` に**足していない**（足すと「中心部が
+  小さくなった」という測定そのものが無意味になる）。2回目の7件も同じ扱い：`r163 #2`（コメント形式）・
+  `r165 #2`（`userTheme` の owner に `keyboard-shortcuts.js` を**意識的に追加**）・`r171`×2 と `r180 ③`
+  （i18n キーは `js/i18n-late.js`）・`r196 ①b`（`tileWarm` のマウントは `js/label-occlusion.js`）・
+  `r198 ②b`（ラベルの STACK は `js/label-occlusion.js`）。
 - `r199 ⑥`（ビルドスタンプ）は**否定形**に置き換え、厳密値のピンは `r200 ⑧` が持つ（#R199 の訂正の継続）。
 
 ### ⑥ できなかったこと・残していること
